% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_data.R
\name{ani_data}
\alias{ani_data}
\title{ani_data}
\usage{
ani_data(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment, provided as an integer}
}
\value{
Dataframe
}
\description{
Get resquest for animal data
}
