% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_data_summary.R
\name{ani_data_summary}
\alias{ani_data_summary}
\title{ani_data_summary}
\usage{
ani_data_summary(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe
}
\description{
Get request for summary animal data
}
