% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticate.R
\name{authenticate}
\alias{authenticate}
\title{authenticate}
\usage{
authenticate(username, password, root_url)
}
\arguments{
\item{username}{User's username address for accound associated with \code{root_url}}

\item{password}{User's password}
}
\description{
This function authenticates a user session. It generates a authorization token
from the user's username and password and then saves the root_url and auth_header to global environment.
Auth_header is the header containing the authorization token.
}
\examples{
\dontrun{authenticate("email@email.com", getPass(), "https://hawcprd.epa.gov")}

}
