% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_data.R
\name{epi_data}
\alias{epi_data}
\title{epi_data}
\usage{
epi_data(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe
}
\description{
Get request for epidemiologic data
}
