% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimeta_data.R
\name{epimeta_data}
\alias{epimeta_data}
\title{epimeta_data}
\usage{
epimeta_data(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe
}
\description{
Get request for meta epidemiologic data
}
