% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_}
\alias{get_}
\title{get_}
\usage{
get_(url, header = auth_header)
}
\arguments{
\item{url}{Full URL for API request}

\item{header}{Header for GET request. If \code{NULL} the default is \code{auth_header}, which contains authorization token}
}
\description{
Basically GET but prints out error message based on status code or formats request response as a dataframe.
}
