% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitro_data.R
\name{invitro_data}
\alias{invitro_data}
\title{invitro_data}
\usage{
invitro_data(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe
}
\description{
Get request for in vitro data
}
