% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_reference_ids.R
\name{lit_reference_ids}
\alias{lit_reference_ids}
\title{lit_reference_ids}
\usage{
lit_reference_ids(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe with variables reference_id, hero_id, pubmed_id
}
\description{
Get request for lit reference ids
}
