% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_reference_tags.R
\name{lit_reference_tags}
\alias{lit_reference_tags}
\title{lit_reference_tags}
\usage{
lit_reference_tags(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe with variables reference_id and tag_id
}
\description{
Get request for lit_reference_tags
}
