% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_references.R
\name{lit_references}
\alias{lit_references}
\title{lit_references}
\usage{
lit_references(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe with a LOT of reference information
}
\description{
Get request for lit references
}
