% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_tags.R
\name{lit_tags}
\alias{lit_tags}
\title{lit_tags}
\usage{
lit_tags(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe with variables id, depth, name, nested_name
}
\description{
Get request for lit tags
}
