% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_data.R
\name{rob_data}
\alias{rob_data}
\title{rob_data}
\usage{
rob_data(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe
}
\description{
Get request for risk of bias data
}
