% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_list.R
\name{visual_list}
\alias{visual_list}
\title{visual_list}
\usage{
visual_list(assessment_id)
}
\arguments{
\item{assessment_id}{HAWC assessment id, provided as an integer}
}
\value{
Dataframe orrrr a list?
}
\description{
Get request for in visual list (?)
}
